requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
PyYAML==6.0.1
scikit-learn==1.4.0
xgboost==1.7.6
